// frontend/src/app/models/gsection.model.ts
export interface Gsection {
    id: String;
    title: String;
    user: String;
    content: String;
    version: String;
    created: Date;
    updated: Date;
    category: String;
    target: String;
    author_id: String;
    author_name: String;
    language: String;
    currency: String;
    price: Number;
    emargin: Number;
    access: Number;
    sold: Number;
}